juicer-backend/requirements.txt
discord.py
requests
aiohttp
fastapi[standard]
uvicorn
python-dotenv